backend/requirements.txt
# Web Framework & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
motor==3.3.2
pymongo==4.6.0

# HTTP Clients
requests==2.31.0
urllib3==2.1.0

# SSE Stream Client
sseclient-py==1.8.0

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
joblib==1.3.2
numpy==1.26.2
pandas==2.1.3

# Geographic calculations
geopy==2.4.1

# Utilities
python-dotenv==1.0.0

lightgbm